Refactored, added checkPath, added Stdout (colors!), map of loggers
The three error levels (Info, Warning, Fatal) are wrappers for Error().
Error will call writeStdout and writeToFile for each log entry. The
stdout will be colored depending on the level. Error() can be called
directly if the dev so chooses. The time layout and path for each logger
can be specified. This allows an application wide logger going to
logs/app and a database logger going to logs/db. The location is the folder
and the filename is the error level.

Plans include to add a timer log that will simply record the time between
start and end and add the duration to a log text.