Option to disable write to file, added timers, info is now cyan
Devs can now disable file writting for a logger. This would be used
if only stdout is needed for dev or similar testing.

Loggers can now have a timer associated with them. Each logger may
only have one timer going at a time. Starting a timer on a logger
without stopping the previous one will reset the timer to the newest
invoke.

Info was too hard to read in blue so it's now cyan. Blue is still
available as a color option.