Moved functions, StopTimer returns string, Location->Path, fixed file
Functions dealing with writing to stdout or file are in write.go.

StopTimer() won't issue an Info log if the string is blank. It now
returns the time elapsed as a string.

logger.Location has been renamed to logger.Path to better reflect
what it does. README is updated to reflect this.

The file property wasn't initialized as true so it wasn't writing
to file by default. It does now.